SIM: drm/i915: Remove modesetting warnings in simulation
The modesetting on the simulator is very minimally supported. Trying to
determine if state is correct is a lost cause. In effect it causes
things to be even slower as all the warnings are printed out.

FIXME: We should remove this if/when the simulator supports it.

v2: Fixup rebase conflict.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>