SIM: drm/i915: Simulator doesn't have all instdone
At least so far, I've never seen the registers return !0, which means
this should impact anything. It's really a safeguard in case in the
future the simulator does decide to return !0.

v2: Rebase on top of the i915_gpu_error.c extraction. I've decided to
leave the gen ordering in natural, the simulator check should still
dtrt.

v3: Move to the end of -internal (Damien)

Signed-off-by: Ben Widawsky <ben@bwidawsk.net> (v1)
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>