SIM: drm/i915: Add simulator's host bridge
Add the host bridge ID used by the simulator. It also gives us an
opportunity to let the rest of the driver know we're running as the
simulator for various workarounds.

We must always do this early as it's the only way we have to detect the
simulator.

Simulation only supports IVB+

v2:
Pull simulation detection out to a separate function (Ben)
Make num_pch_plls 0 (Ben)

v3: Move the settings the new function to the top to reduce churn in a
later patch (Ben)

v4: Kill dev_priv->num_pch_pll assignment, that's now handled
elseplace.

v5: Move to the end of drm-intel-internal (Damien)

Signed-off-by: Ben Widawsky <ben@bwidawsk.net> (v3)
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>