SIM: drm/i915/chv: VLV and CHV don't have a PCH
So intel_pch_detect() shouldn't try to find it when in simulation mode.

As a note about the original check for >= GEN5: we don't have simulator
support for these early gens. While the original patch was indeed
correct to disable PCH checking for < GEN5, it's not needed here and
made the condition check really ugly due to 80 character columns.

v2: Moved to new isolated detect function (Ben)

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com> (v1)
[bwidawsk: Dropped IS_GEN5 check, and modified commit msg]
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>