SIM: drm/i915/bdw: Add simulator PCI ids
Over the course of rebasing the internal repository, it seems the
simulator IDs have been added to the main BDW macro. These have been
separated out of the external patches. Here we re-enable for simulation
support.

It is still my hope to one day not use special PCI ids for the simulator
- I am just lazy.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>