drm/i915/skl: GuC scheduler setup
The GuC has scheduling capabilities, and can take over for the driver in
certain cases. The scheduler support requires setting up an object which
can be shared between the driver and firmware (the context pool). In
addition, there are several fields which are meant to be passed from the
driver to the firmware which are missing from the patch.

This patch should represent the minimal code needed to bring up
firmwares with scheduling capability. Platform specific needs will be
addressed in subsequent patches.

v2: Use 1024 contexts, which is 69 pages in size (Vinit).

v3: Allocate number of pages * PAGE_SIZE, not just number of pages
(Vinit)
[v4 by Vinit]: Don't unreference unallocated object (Jon/Brad)
v5: Remove extra SOFT_SCRATCH1 definition. Note sure when it got in
there.

Signed-off-by: Vinit Azad <vinit.azad@intel.com> (v4)
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Alex Dai <yu.dai@intel.com>
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>