!UPSTREAM drm/i915/GuC: Support BDW GuC
GuC is not POR on Broadwell. It can be used for prototyping however.
This patch which we can decide to later upstream if the relevant fuses
are programmed for shipping BDW, simply flips the switch for BDW as we
just did for GEN9.

v2: Change IS_CHV || IS_BDW check to IS_GEN8 (Ben)

v3: Rebase on top of the HAS_FDI() removal (damien)

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Alex Dai <yu.dai@intel.com>
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>