drivers/misc: Baytrail SoC platform clock support
The Baytrail SoC provides 6 platform clocks which can run either
at 25MHz (xtal) or 19.2 MHz (PLL) and with an on/off control.

These clocks are not exposed by ACPI for now and given the
simplicity of the controls there is no real point in porting
this code to the common clock framework.

ASoC machine drivers for Baytrail are the main users of this
driver. The CLK3 is used as an audio MCLK used by the
audio codec as its PLL input, which results in better audio
quality than if the I2S bitclock is used for the PLL input.