ASoC: Intel: bytct-rt5640: add AIF_SWAP quirk
Some Baytrail devices are wired with a SSP0-AIF2 connection
(typically all Baytrail-CR devices) instead of the regular
SSP2-AIF1 connection. This patch adds a quirk to swap AIF1 and AIF2.

This quirk can only be used on regular hardware when using the
modified SST firmware which swaps SSP2 and SSP0. To use the same settings
on the codec, the controls "DAI select" and "SDI select" to 1 re-swap
AIF1 and AIF2 when this quirk is used.