ASoC: rt5645: add fallback case for jack detection support
Commit 78f5605c0329 ('ASoC: rt5645: cleanup DMI matching code')
did a lot of useful cleanups, but the logic is a bit flawed:
if there is no DMI match on an ACPI machine, the code reads non-existent
properties.

Re-add previous code that checked for the existence of those properties
and a fallback mode to enable jd-mode3 if there is no pdata, device property
or quirk. This will help avoid having to add new DMI quirks when only
the jd-mode is needed for headset playback.

Cc: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>