pinctrl: cherryview: Adjust interrupt line mapping
As rightfully stated in the comment of INVALID_HWIRQ() macro
"Contrary to Linux irqs, for hardware irqs the irq number 0 is valid".

This is happens when got an interrupt from the source without handler
assigned which effectively maps the IRQ to offset 0, due to kzalloc().

In the result we have wrong interpretation in /proc interrupts:
	115:   ... chv-gpio    0
	171:   ... chv-gpio    0

where the source apparently is not a pin 0 of certain GPIO community.

Fix this by using obviously invalid HW IRQ number as default for no IRQ
assigned.

For backporting the fix either commit d17bf24e6952 or local define of
INVALID_HWIRQ constant should be applied.

Reported-by: Guenter Roeck <linux@roeck-us.net>
Cc: Mika Westerberg <mika.westerberg@linux.intel.com>
Fixes: 6e08d6bbebeb ("pinctrl: Add Intel Cherryview/Braswell pin controller support")
Requires: d17bf24e6952 ("genirq: Add a new generic IPI reservation code to irq core")
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>