ASoC: Intel: boards: add implementation for PCM512x clock driver
Very simple driver which just registers a clock, and flips the
relevant GPIOs by directly accessing the regmap passed as platform
data.

The clock cannot be changed on the fly after it is prepared, the
machine driver should unprepare the clock when in idle.

FIXME:

1. does this code below in an Intel-specific directory? It's difficult
to get things merged in the driver/clk side of things.

2. test on RPI platform, this should be compatible with the existing
Hifiberry overlays.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>