ASoC: pcm512x: expose 6 GPIOs
The GPIOs are used e.g. on HifiBerry DAC+ HATs to control the LED
(GPIO3) and the choice of the 44.1 (GPIO6) or 48 (GPIO3) kHz
oscillator (when present).

Enable gpio_regmap to get/set values and get/set directions. Tested
with GPIO_LIB from sys/class/gpio, the LED turns on/off as desired.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>

transition to gpio_regmap

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>