ASoC: pcm512x: expose 6 GPIOs
The GPIOs are used e.g. on HifiBerry DAC+ HATs to control the LED
(GPIO3) and the choice of the 44.1 (GPIO6) or 48 (GPIO3) kHz
oscillator (when present).

Enable basic gpio_chip to get/set values and get/set
directions. Tested with GPIO_LIB from sys/class/gpio, the LED turns
on/off as desired.

A lookup table is added to let the clock driver request the
gpios. dev_id is not set, so gpiod_get() will have to use NULL device

TODO:

2. race condition between pcm512x driver exposing gpio chip used by
clk driver using gpios to provide clocks.

3. checking if the clocks actually work is done by checking the
PCM512x_RATE_DET_4 register (after typically a 2ms sleep), so maybe
the regmap does need to be passed to the clock driver anyways.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>