ASoC: Intel: Skylake: move i915 component init to initial probe
The current flow makes no sense. The i915 component is initialized in
the probe workqueue but released in the initial probe. There is no
reason to defer this anyway, the sooner this binding is established
the better: this patch solves timing issues observed on Dell XPS13
9350 and ApolloLake devices, where HDMI/iDISP was not detected and
invalid configurations passed to the machine driver.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>