ASoC: core: fixes to show new check for probe_defer impacts BE overrides
With the upstream changes, the tests on the dailink platform names are
done before we even have a chance to override them. Moving the call of
soc_check_tplg_fes() from soc_instantiate_card() to
soc_register_card() partly solves the problem. This can be shown with
the PLB_FIX1 define.

There is however a second problem where we seem to have a memory leak
in snd_soc_init_platform(), the use of the legacy_plaform field makes
no sense. this can be shown with the PLB_FIX2 define.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>